Simple Budget Windows Installation v2.1.5
Run this to install Simple Budget in Windows.  The installer will ask for a directory to install the files to, and run simplebudget.exe in that directory.  Simple Budget will create a folder named simplebudget for data files under the users profile directory the first time simplebudget.exe is ran.